+++ atlasdb-client/src/main/java/com/palantir/atlasdb/keyvalue/impl/DualWriteKeyValueService.java
@@ -131,15 +131,10 @@ public class DualWriteKeyValueService implements KeyValueService {
     }
 
     @Override
-    public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell) {
-        delegate1.deleteAllTimestamps(tableRef, maxTimestampExclusiveByCell);
-        delegate2.deleteAllTimestamps(tableRef, maxTimestampExclusiveByCell);
-    }
-
-    @Override
-    public void deleteAllTimestampsAndSentinels(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell) {
-        delegate1.deleteAllTimestampsAndSentinels(tableRef, maxTimestampExclusiveByCell);
-        delegate2.deleteAllTimestampsAndSentinels(tableRef, maxTimestampExclusiveByCell);
+    public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell,
+            boolean deleteSentinels) {
+        delegate1.deleteAllTimestamps(tableRef, maxTimestampExclusiveByCell, deleteSentinels);
+        delegate2.deleteAllTimestamps(tableRef, maxTimestampExclusiveByCell, deleteSentinels);
     }
 
     @Override