+++ atlasdb-client/src/main/java/com/palantir/atlasdb/keyvalue/impl/ProfilingKeyValueService.java
@@ -179,15 +179,10 @@ public final class ProfilingKeyValueService implements KeyValueService {
     }
 
     @Override
-    public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell) {
-        maybeLog(() -> delegate.deleteAllTimestamps(tableRef, maxTimestampExclusiveByCell),
+    public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell,
+            boolean deleteSentinels) {
+        maybeLog(() -> delegate.deleteAllTimestamps(tableRef, maxTimestampExclusiveByCell, deleteSentinels),
                 logTimeAndTable("deleteAllTimestamps", tableRef));
-    }
-
-    @Override
-    public void deleteAllTimestampsAndSentinels(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell) {
-        maybeLog(() -> delegate.deleteAllTimestampsAndSentinels(tableRef, maxTimestampExclusiveByCell),
-                logTimeAndTable("deleteAllTimestampsAndSentinels", tableRef));
     }
 
     @Override